Refactor shared scatter test functionality into base class.
