+++ glue_ar/common/tests/test_scatter.py
@@ -4,10 +4,20 @@ from glue_vispy_viewers.scatter.qt.scatter_viewer import VispyScatterViewer
 from itertools import product
 from math import sqrt
 from numpy import array, array_equal, nan, ones
+from os import remove
 import pytest
-from typing import cast
+from random import random, randint, seed
+from typing import cast, Dict, Tuple, Type, Union
+
+from glue.core.state_objects import State
+from glue.viewers.common.viewer import Viewer
+from glue_jupyter import JupyterApplication
+from glue_jupyter.ipyvolume.scatter import IpyvolumeScatterView
+from glue_qt.app import GlueApplication
 
 from glue_ar.common.scatter import scatter_layer_mask
+from glue_ar.common.scatter_export_options import ARIpyvolumeScatterExportOptions, ARVispyScatterExportOptions
+from glue_ar.utils import export_label_for_layer
 
 
 @pytest.fixture
@@ -82,3 +92,82 @@ def test_scatter_mask_bounds(scatter_mask_data, clip, size, color):
         assert array_equal(expected, mask)
     else:
         assert mask is None
+
+
+@pytest.mark.parametrize("app_type,viewer_type", (("qt", "vispy"), ("jupyter", "vispy"), ("jupyter", "ipyvolume")))
+class BaseScatterTest:
+
+    @pytest.fixture(scope='function', autouse=True)
+    def setup_method(self, app_type: str, viewer_type: str):
+
+        # Setup
+        self.app_type = app_type
+        self.viewer_type = viewer_type
+        seed(186)
+        self.n = 40
+        x1 = [random() * 5 for _ in range(self.n)]
+        y1 = [random() for _ in range(self.n)]
+        z1 = [randint(1, 30) for _ in range(self.n)]
+        self.data1 = Data(x=x1, y=y1, z=z1, label="Scatter Data 1")
+        self.data1.style.color = "#fedcba"
+        self.app = self._create_application(app_type)
+        self.app.data_collection.append(self.data1)
+        self.viewer: Viewer = self.app.new_data_viewer(self._viewer_class(viewer_type))
+        self.viewer.add_data(self.data1)
+
+        x2 = [random() * 7 for _ in range(self.n)]
+        y2 = [randint(100, 200) for _ in range(self.n)]
+        z2 = [random() for _ in range(self.n)]
+        self.data2 = Data(x=x2, y=y2, z=z2, label="Scatter Data 2")
+
+        self.viewer.state.x_att = self.data1.id['x']
+        self.viewer.state.y_att = self.data1.id['y']
+        self.viewer.state.z_att = self.data1.id['z']
+        self.state_dictionary = self._basic_state_dictionary(viewer_type)
+
+        yield
+
+        # Teardown
+        if getattr(self, "tmpfile", None) is not None:
+            self.tmpfile.close()
+            remove(self.tmpfile.name)
+        if hasattr(self.viewer, "close"):
+            self.viewer.close(warn=False)
+        self.viewer = None
+        if hasattr(self.app, 'close'):
+            self.app.close()
+        self.app = None
+
+    def _create_application(self, app_type: str) -> Union[GlueApplication, JupyterApplication]:
+        if app_type == "qt":
+            return GlueApplication()
+        elif app_type == "jupyter":
+            return JupyterApplication()
+        else:
+            raise ValueError("Application type should be either qt or jupyter")
+
+    def _viewer_class(self, viewer_type: str) -> Union[Type[VispyScatterViewer], Type[IpyvolumeScatterView]]:
+        if viewer_type == "vispy":
+            return VispyScatterViewer
+        elif viewer_type == "ipyvolume":
+            return IpyvolumeScatterView
+        else:
+            raise ValueError("Viewer type should be either vispy or ipyvolume")
+
+    def _basic_state_dictionary(self, viewer_type: str) -> Dict[str, Tuple[str, State]]:
+        if viewer_type == "vispy":
+            state_maker = lambda: ARVispyScatterExportOptions(theta_resolution=15,
+                                                              phi_resolution=15)
+        elif viewer_type == "ipyvolume":
+            state_maker = ARIpyvolumeScatterExportOptions
+        else:
+            raise ValueError("Viewer type should be either vispy or ipyvolume")
+
+        return {
+            export_label_for_layer(layer): ("Scatter", state_maker())
+            for layer in self.viewer.layers
+        }
+
+    def _export_state_class(self, viewer_type: str):
+        return ARIpyvolumeScatterExportOptions if viewer_type == "ipyvolume" else ARVispyScatterExportOptions
+
