+++ glue_ar/common/tests/test_scatter_gltf.py
@@ -1,99 +1,17 @@
-from os import remove
-from random import random, randint, seed
 from tempfile import NamedTemporaryFile
-from typing import Dict, Tuple, Type, Union
 
 from gltflib import AccessorType, AlphaMode, BufferTarget, ComponentType, GLTFModel
 from gltflib.gltf import GLTF
-from glue.core import Data
-from glue.core.state_objects import State
-from glue.viewers.common.viewer import Viewer
-from glue_jupyter import JupyterApplication
-from glue_jupyter.ipyvolume.scatter import IpyvolumeScatterView
-from glue_qt.app import GlueApplication
-from glue_vispy_viewers.scatter.qt.scatter_viewer import VispyScatterViewer
-import pytest
 
 from glue_ar.common.export import export_viewer
 from glue_ar.common.scatter_export_options import ARIpyvolumeScatterExportOptions, ARVispyScatterExportOptions
 from glue_ar.common.shapes import sphere_points_count, sphere_triangles, sphere_triangles_count
 from glue_ar.common.tests.gltf_helpers import count_indices, count_vertices, unpack_vertices
+from glue_ar.common.tests.test_scatter import BaseScatterTest
 from glue_ar.utils import export_label_for_layer, hex_to_components, layers_to_export, mask_for_bounds, xyz_bounds, xyz_for_layer
 
 
-@pytest.mark.parametrize("app_type,viewer_type", (("qt", "vispy"), ("jupyter", "vispy"), ("jupyter", "ipyvolume")))
-class TestScatterGLTF:
-
-    @pytest.fixture(scope='function', autouse=True)
-    def setup_and_teardown(self, app_type: str, viewer_type: str):
-
-        # Setup
-        seed(1374)
-        self.n = 80
-        x1 = [random() * 5 for _ in range(self.n)]
-        y1 = [random() for _ in range(self.n)]
-        z1 = [randint(1, 30) for _ in range(self.n)]
-        self.data1 = Data(x=x1, y=y1, z=z1, label="Scatter Data 1")
-        self.data1.style.color = "#fedcba"
-        self.app = self._create_application(app_type)
-        self.app.data_collection.append(self.data1)
-        self.viewer_type = viewer_type
-        self.viewer: Viewer = self.app.new_data_viewer(self._viewer_class(viewer_type))
-        self.viewer.add_data(self.data1)
-
-        x2 = [random() * 7 for _ in range(self.n)]
-        y2 = [randint(100, 200) for _ in range(self.n)]
-        z2 = [random() for _ in range(self.n)]
-        self.data2 = Data(x=x2, y=y2, z=z2, label="Scatter Data 2")
-
-        self.viewer.state.x_att = self.data1.id['x']
-        self.viewer.state.y_att = self.data1.id['y']
-        self.viewer.state.z_att = self.data1.id['z']
-        self.state_dictionary = self._basic_state_dictionary(viewer_type)
-
-        yield
-
-        # Teardown
-        if getattr(self, "tmpfile", None) is not None:
-            self.tmpfile.close()
-            remove(self.tmpfile.name)
-        if hasattr(self.viewer, "close"):
-            self.viewer.close(warn=False)
-        self.viewer = None
-        if hasattr(self.app, 'close'):
-            self.app.close()
-        self.app = None
-
-    def _create_application(self, app_type: str) -> Union[GlueApplication, JupyterApplication]:
-        if app_type == "qt":
-            return GlueApplication()
-        elif app_type == "jupyter":
-            return JupyterApplication()
-        else:
-            raise ValueError("Application type should be either qt or jupyter")
-
-    def _viewer_class(self, viewer_type: str) -> Union[Type[VispyScatterViewer], Type[IpyvolumeScatterView]]:
-        if viewer_type == "vispy":
-            return VispyScatterViewer
-        elif viewer_type == "ipyvolume":
-            return IpyvolumeScatterView
-        else:
-            raise ValueError("Viewer type should be either vispy or ipyvolume")
-
-    def _basic_state_dictionary(self, viewer_type: str) -> Dict[str, Tuple[str, State]]:
-        if viewer_type == "vispy":
-            state_maker = lambda: ARVispyScatterExportOptions(theta_resolution=15,
-                                                              phi_resolution=15)
-        elif viewer_type == "ipyvolume":
-            state_maker = ARIpyvolumeScatterExportOptions
-        else:
-            raise ValueError("Viewer type should be either vispy or ipyvolume")
-
-        return {
-            export_label_for_layer(layer): ("Scatter", state_maker())
-            for layer in self.viewer.layers
-        }
-
+class TestScatterGLTF(BaseScatterTest):
 
     # TODO: How can we test the properties of compressed files?
     def test_basic_export(self):
@@ -124,7 +42,7 @@ class TestScatterGLTF:
         label = export_label_for_layer(layer)
         method, options = self.state_dictionary[label]
         assert method == "Scatter"
-        export_state_cls = ARIpyvolumeScatterExportOptions if self.viewer_type == "ipyvolume" else ARVispyScatterExportOptions
+        export_state_cls = self._export_state_class(self.viewer_type)
         assert isinstance(options, export_state_cls)
         color_components = [c / 256 for c in hex_to_components("#fedcba")] + [layer.state.alpha]
         assert material.alphaMode == AlphaMode.BLEND.value
