+++ glue_ar/common/tests/test_scatter_usd.py
@@ -1,97 +1,16 @@
-from os import remove
-from random import random, randint, seed
 from tempfile import NamedTemporaryFile
-from typing import Dict, Tuple, Type, Union
 
-from glue.core import Data
-from glue.core.state_objects import State
-from glue.viewers.common.viewer import Viewer
-from glue_jupyter import JupyterApplication
-from glue_jupyter.ipyvolume.scatter import IpyvolumeScatterView
-from glue_qt.app import GlueApplication
-from glue_vispy_viewers.scatter.qt.scatter_viewer import VispyScatterViewer
 from pxr import Sdf, Usd
-import pytest
 
 from glue_ar.common.export import export_viewer
-from glue_ar.common.scatter import IPYVOLUME_POINTS_GETTERS, radius_for_scatter_layer
-from glue_ar.common.scatter_export_options import ARIpyvolumeScatterExportOptions, ARVispyScatterExportOptions
-from glue_ar.common.shapes import sphere_points, sphere_points_count, sphere_triangles_count
+from glue_ar.common.shapes import sphere_points_count, sphere_triangles_count
 from glue_ar.usd_utils import material_for_mesh
 from glue_ar.utils import export_label_for_layer, hex_to_components, iterator_count, layers_to_export, xyz_bounds
 
+from glue_ar.common.tests.test_scatter import BaseScatterTest
 
-@pytest.mark.parametrize("app_type,viewer_type", (("qt", "vispy"), ("jupyter", "vispy"), ("jupyter", "ipyvolume")))
-class TestVispyScatterUSD:
 
-    @pytest.fixture(scope='function', autouse=True)
-    def setup_method(self, app_type: str, viewer_type: str):
-
-        # Setup
-        seed(186)
-        self.n = 40
-        x1 = [random() * 5 for _ in range(self.n)]
-        y1 = [random() for _ in range(self.n)]
-        z1 = [randint(1, 30) for _ in range(self.n)]
-        self.data1 = Data(x=x1, y=y1, z=z1, label="Scatter Data 1")
-        self.data1.style.color = "#fedcba"
-        self.app = self._create_application(app_type)
-        self.app.data_collection.append(self.data1)
-        self.viewer: Viewer = self.app.new_data_viewer(self._viewer_class(viewer_type))
-        self.viewer.add_data(self.data1)
-
-        x2 = [random() * 7 for _ in range(self.n)]
-        y2 = [randint(100, 200) for _ in range(self.n)]
-        z2 = [random() for _ in range(self.n)]
-        self.data2 = Data(x=x2, y=y2, z=z2, label="Scatter Data 2")
-
-        self.viewer.state.x_att = self.data1.id['x']
-        self.viewer.state.y_att = self.data1.id['y']
-        self.viewer.state.z_att = self.data1.id['z']
-        self.state_dictionary = self._basic_state_dictionary(viewer_type)
-
-        yield
-
-        # Teardown
-        if getattr(self, "tmpfile", None) is not None:
-            self.tmpfile.close()
-            remove(self.tmpfile.name)
-        if hasattr(self.viewer, "close"):
-            self.viewer.close(warn=False)
-        self.viewer = None
-        if hasattr(self.app, 'close'):
-            self.app.close()
-        self.app = None
-
-    def _create_application(self, app_type: str) -> Union[GlueApplication, JupyterApplication]:
-        if app_type == "qt":
-            return GlueApplication()
-        elif app_type == "jupyter":
-            return JupyterApplication()
-        else:
-            raise ValueError("Application type should be either qt or jupyter")
-
-    def _viewer_class(self, viewer_type: str) -> Union[Type[VispyScatterViewer], Type[IpyvolumeScatterView]]:
-        if viewer_type == "vispy":
-            return VispyScatterViewer
-        elif viewer_type == "ipyvolume":
-            return IpyvolumeScatterView
-        else:
-            raise ValueError("Viewer type should be either vispy or ipyvolume")
-
-    def _basic_state_dictionary(self, viewer_type: str) -> Dict[str, Tuple[str, State]]:
-        if viewer_type == "vispy":
-            state_maker = lambda: ARVispyScatterExportOptions(theta_resolution=15,
-                                                              phi_resolution=15)
-        elif viewer_type == "ipyvolume":
-            state_maker = ARIpyvolumeScatterExportOptions
-        else:
-            raise ValueError("Viewer type should be either vispy or ipyvolume")
-
-        return {
-            export_label_for_layer(layer): ("Scatter", state_maker())
-            for layer in self.viewer.layers
-        }
+class TestVispyScatterUSD(BaseScatterTest):
 
     def test_basic_export(self):
         bounds = xyz_bounds(self.viewer.state, with_resolution=False)
